+++ core/src/emulator/io/fdc/fdc.h
@@ -5,13 +5,13 @@
 // FDC-specific constants and functionality defined
 
 /// region <Constants>
-static constexpr size_t FDD_RPS = 5;                    // FDD rotation speed
+static constexpr uint8_t FDD_RPS = 5;                   // FDD rotation speed
 static constexpr size_t MAX_TRACK_LEN = 6250;           //
 static constexpr size_t MAX_SECTOR_DATA_LEN = 6144;     //
-static constexpr size_t MAX_CYLINDERS = 86;             // Don't load images with track number exceeding this threshold
-static constexpr size_t MAX_SIDES = 2;                  // There can be single side disks
-static constexpr size_t MAX_PHYSICAL_CYLINDER = 86;     // FDC won't perform seek commands beyond this cylinder
-static constexpr size_t MAX_SECTOR = 32;               // Max sectors per track
+static constexpr uint8_t MAX_CYLINDERS = 86;            // Don't load images with track number exceeding this threshold
+static constexpr uint8_t MAX_SIDES = 2;                 // There can be single side disks
+static constexpr uint8_t MAX_PHYSICAL_CYLINDER = 86;    // FDC won't perform seek commands beyond this cylinder
+static constexpr uint8_t MAX_SECTOR = 32;               // Max sectors per track
 /// endregion </Constants>
 
 /// region <Types>
